MAINT: stats: migrate circular statistics functions to _circstats.py
